Module Pbrt_yojson

Protobuf JSON encoding runtime

module E : sig ... end

All exception which could be raised by the generated JSON encoder and decode function

Helper module for the generated code for common functionality

val string : Yojson.Basic.t -> string -> string -> string

val float : Yojson.Basic.t -> string -> string -> float

val int32 : Yojson.Basic.t -> string -> string -> int32

val int64 : Yojson.Basic.t -> string -> string -> int64

val int : Yojson.Basic.t -> string -> string -> int

val bool : Yojson.Basic.t -> string -> string -> bool

val bytes : Yojson.Basic.t -> string -> string -> bytes

val unit : Yojson.Basic.t -> string -> string -> unit

val make_bool : bool -> Yojson.Basic.t

val make_int : int -> Yojson.Basic.t

val make_float : float -> Yojson.Basic.t

val make_string : string -> Yojson.Basic.t

val make_bytes : bytes -> Yojson.Basic.t

val make_unit : unit -> Yojson.Basic.t

val make_list : Yojson.Basic.t list -> Yojson.Basic.t